Primera Parte pantalla feed con repo casi hecho

+++ app/src/main/java/com/example/soundlink/core/data/dto/RegisterRequest.kt
@@ -4,5 +4,5 @@ data class RegisterRequest(
     val name: String,
     val email: String,
     val password: String,
-    val age: Long
+    val age: Int
 )
